refine #298:
In case user is moving fast (passenger): distance is now the minimum distance between quest geometry and line between location when the form was opened and when the user clicked "okay"

Correct distance calculation into meters. Before, it was simply wrong (calculated bigger distances the further you were from the equator).

Change layout of dialog again: Include the reason why the question is asked in the first place (=the rough measured distance to the quest) and make the note about the stars less prominent.

Concession: the app offers the "I am there option" for quests that are up to 250 away but only does not ask for less than 50 away.

# Conflicts:
#	app/src/main/res/values/strings.xml