Correct wording of tag/keys
I also always get that wrong, but in "OSM-speak", the whole thing is a tag, and the other stuff is just as a programmer would expect, i.e. 

Tag -> `tourism=museum` = `key=value`

So one usually adds/modifies the value, or, in general the tag. A key is rather not modified.