Opening Hours: Expect Swedish opening hours to differ on Saturdays
While a lot of shops in a big city in Sweden are open all days of
the week, Monday to Friday tend to have the same opening hours,
but shorter opening hours on Saturdays and Sundays.
This means that when setting the opening hours in StreetComplete,
you always have to uncheck Saturday first.

This commit fixes it to be Monday-Friday by default.