Avoid possible user mistake by only allowing to finish note with keyboard closed (#1039)
On composing a note, first close the keyboard on click on OK. Only when keyboard is closed will a click on OK lead to the note being saved. This enables to user to always see the exact position of the note before he saves it, as the note position (necessarily) shifts when rotating the phone.