Add British English (en-GB) translation.
Note the translation contains *all* the strings, even the ones
that are identical to the original US English source strings.
We do this to ensure that we always know which strings have actually
been considered for ‘translation’ into British English.
Along with the other changes in this branch, this fixes #914.