remove the option to add building:description (for now)
Because in case a building is "just a building, can't say what for exactly", the user will also not be able to provide a meaningful "concise description of its purpose", so the meaning of building:description defeats itself - it does not seem like a solution for that problem.
If the user CAN provide a description, then a note should be created so that it can be tagged later properly.