Add a meaningful user agent to request for map tiles
This will add a User-Agent headder in the form of
"StreetComplete 7.0 / okhttp/3.5.0" instant of the
default "okhttp/3.5.0" to requests for tiles.

Unfortunately it seems to be necessary to copy the onRequest method from
tangram, as a change in the middle is necessary.

This change is mostly for debugging on the tile Server.