fix possible crash when opening the answer form for an element that has been deleted already
During conflict resolution on upload, the element to which a quest refers to is updated from the OSM server. Since right about anything could have changed, it must be rechecked which quests are applicable to this element. And in case the element has been deleted, all the quests on that element need to be removed.
Before, such check was only done for the quest currently being uploaded.