make questions signs more visible in the sidewalk and cycleway quest
only one tester was solving this quest type (as the first ever in her case)
and it went like this

- select this quest
- see road with no sidewalk
- select no sidewalk
- tap outside quest

More visible question signs may help to avoid this mistake as it will be more
obvious that more remain to solve this. Current ones
are not clearly visible for someone with poor eyesignt or in a glaring sun.