articles fixes
As "OpenStreetMap" is used as a proper noun, adding `the` in front of it is incorrect. `the` *would* be correct for "the OpenStreetMap database" or "the OpenStreetMap project".
Change "the" to "its": AFAIK, StreetComplete is the only project to use these map styles.
Change "the" to "a": Many quests have a different simple question depending on whether or not the object is named, or due to the type of object.