exclude ruined buildings from quests
Such buildings are often not really answerable and value of answers will be low. Large number of questions will end in failure to reply.

Some ruins were built was reasons that are not clear now, a roof may be caved in making the roof shape question unanswerable (and not really fixable by tagging it), building levels are weird at the best (is 0 correct in case of completely missing roof) etc.

Though I admit that the real reason is that at least in some places many object could have historic=* tag (that is already filtering out objects in building type quest), but historic key is untagged. Many of such objects have ruins=yes.

fixes #1541