completely show bottom sheet when keyboard comes up
keyboard-is-up detection a bit hacky, but there is no other way. When the bottom sheet was not up completely, the keyboard would not push up the content of the bottom sheet. Not sure why this happens, the whole bottom sheet is a bit of a black box to me. This fixes that.