do not allow backup
backup has some security implications. Plus, the local data of StreetComplete will almost all the time be above the 25MB limit, so backup does not work most of the time anyway. To backup downloaded quests makes no sense. The only thing that makes sense would be not-uploaded-yet answers, but they should not be stored for a lengthy period anyway plus we wouldn't want these answers to be spread over all of the user's devices (cause then there would be upload conflicts)