Position gps button to left of zoom controls
When gps is enabled and the button disappears, it's awkward to have an
empty space below the zoom buttons. This moves the gps button to the
side and shifts the zoom buttons down, forming a .: shape. This way,
when the gps button disappears, the zoom buttons sit nicely at the
bottom.

This is a proof of concept implementation; it has an awkward interaction
with the "off screen" location pin, since the wrapping LinearLayout is
bigger than the visible buttons, so the pin is shown prematurely.