Hide gps button only when following position
This is a somewhat naive implementation and I have not confirmed that it
has the correct behavior in all edge cases. Still, it should be good
enough to play around with and see what it feels like.

It also exposes a previously-hidden bug — the map will snap, not scroll,
to position if there is any momentum when you press the location button.