Don't follow location on tapping self-marker/pin
Instead, just scroll the view to be centered on the current location.

I'm pretty sure this makes sense for when you tap on the pin, when your
current location is off screen. I'm less sure what makes sense for
tapping on the marker when your location is on screen (if anything).
But for now they both do this.