Remove centering affect on tapping location marker
It makes sense for the location pin (when you are off screen), but the
marker is harder to tap, and displays like the rest of the map, not a
pin — there's no visual indication that it should be interactive.

Left the callback in place in case we want to use it in the future.