Add an aerial layer to VectorTileProvider
Just change the data structure and fix the build, same functionality.

The base and aerial layers are hard coded, which has a few downsides:

- This is a rather inflexible data structure -- if we wanted to add
  additional layers later, it would require another change.

- There is little separation of concerns. For example, the downloader
  will need logic for handling both layers (which will also need to be
  modified if another layer is added).

A better interface would be like this:

    data class MapScene(
        val sceneFilePath: String,
        val sources: List<TileSource>
    )

`sceneFilePath` would link to the actual yaml file (as opposed to just
the folder), and a different `MapScene` would be instantiated depending
on which theme & view was chosen. The downloader would receive the
current scene and simply download all layers (each of which would hold
info about how long it should be cached), rather than checking the
preferences and making the decision about which layers to download,
itself.

However, this interface, without much abstraction, is more flexible. If
a change to the logic needs to be made, it can be made right there in
the downloader, rather than requiring an interface change and tweaks to
multiple components. Then, after the feature has been released for a
while and we've confirmed our assumptions about what information should
or shouldn't be hidden behind an abstraction, I can do another refactor
into a cleaner interface.