do not tag foot=use_sidepath when user answered that sidewalk is separate (fixes  #2895)
foot=use_sidepath is a legal category and while it is usually implied (in most legislations) that if there is a sidewalk (that has been separately mapped), the road itself may not be used by pedestrians (cause they shall use the sidewalk), we don't know if this applies for all legislations.
Apparently this is f.e. not the case for the UK. So, it is safer to leave out this tagging.