Fix missing satellite imagery at high zoom levels when offline
When downloading tiles for offline, the esri TileSource's maxZoom was
arbitrarily set to 16, since that's what the jawg tiles were set to.

According to the [Tangram docs] the map should "overzoom" when zooming
in past the max_zoom layer. However, when tiles were downloaded, then
the phone went offline, zooming past 16 would result in a black
background, instead (missing tiles).

The problem was that Tangram chooses its max_zoom based on the yaml
scene file, not the TileSource declared on kotlin-side. It was not
specified, and the default max_zoom is 18. So, Tangram was trying to
show tiles at zoom levels 17 and 18 -- and failing, of course, since the
phone is offline and they were not downloaded ahead of time.

The fix: make sure the two declarations are in sync.

According to the [ESRI geojson], they support up to zoom level 22.
However, zoom levels 18+ were unavailable at the one location I tested,
so I've gone with 17 for now. It would be good to do more testing to
figure out the highest reliably-present zoom level, and how much data it
uses, and adjust this number if needed.

[Tangram docs]: https://tangrams.readthedocs.io/en/master/Syntax-Reference/sources/#max_zoom
[ESRI geojson]: https://raw.githubusercontent.com/osmlab/editor-layer-index/74501812b31cca1c3505d7e942c5188c63adf5b2/sources/world/EsriImagery.geojson