Convert MapTilesDownloader to flows to fix stats
We had a classic concurrency bug: the stats were being incremented from
different coroutines, which can overwrite each other mid-flight, causing
data to be lost.

Using flows, the stats tracking is limited to a single coroutine, so we
avoid this problem. Flow will also be easier to combine, when we have
multiple data sources.