Remove name from vacant shop when it's no longer vacant.
Vacant shops sometimes have the `name` key still set to help make it identifiable when the house number can't be determined in person. (As shops often don't care about removing signage when they close for good)

When answering this quest on such an element, the name key is often retained and the data is left bad.

I'll provide an example:

A bar near me closed down ("The Black Cat") but left their signage up. I left the name tag on the element when I set "shop=vacant" as it's in a dense part of town where it'll be hard to know what element StreetComplete might be asking for otherwise. A new establishment opened in its place but StreetComplete never asked me for the name as the old bars name wasn't removed from the "shop=vacant" element. 

Whether or not shop=vacant or disused:shop=* elements should have names, 30% of them do, so it's worth covering the edge case :)