Fix team mode dialog text clipping on small devices
Works in English. In a German auto-translation, there is still a little
clipping on a *tiny* device (Nexus S, 4", 480x800px, hdpi) when the
keyboard is open, but never when it is closed.

wrap_content is to prevent truncation.
android:layout_gravity is to re-center the text, which was starting low.