Make tile download stats fully accurate
The stats were being incremented from different coroutines, which can
overwrite each other mid-flight, causing data to be lost.

This is pretty rare. In testing:

- I was only able to trigger the bug on **cached tiles**.
- I was never able to produce more than 1 conflict per download.
- I only observed the bug with ESRI tiles.
    - I believe this was just about volume: the ESRI tiles were using a
      higher zoom, so there were around 4x more of them.
    - The way I refactored the code, the ESRI download did not start
      until after the Jawg download completed; if they were mixed then
      perhaps I'd have seen a conflict in the Jawg stats.
- How often I was able to trigger the bug depended on how many tiles
  were downloaded at a time.
    - When zoomed all the way out (around 250 ESRI tiles), I saw a
      conflict 80% of the time.
    - The fewest number of tiles that produced a conflict was 115.
      (Jawg tiles never went above ~80)

So, it's not important to fix just for stats. However, this protects us
from heisenbugs if we ever hook up another system to this code.

This is just the fix, a small refactor will come in the next commit.