Reverse direction of history list
This makes the most recent edit start at the bottom, so when you tap the
undo button, you don't need to move your hand far to undo the most
recent edit.

It would be nice to have a slightly larger bottom padding, so the button
lines up perfectly and it's just a double-tap.