Add bottomPadding on the edit drawer the right way
The initial position of the most recent edit in the history should line
up with the undo button (above the attribution). That way, when you tap
the undo button, the most recent edit appears directly under your
finger, rather than slightly below it.

Adding this offset as a margin causes funky behavior with items
disappearing to early as your scroll. Adding padding in the xml didn't
work because we override the padding from Kotlin. We'd still prefer to
define the extra padding from XML, though, not Kotlin (who's going to
remember to change that when updating views?). So instead, store the
initial bottom padding and respect that when overriding the padding.