Allow duplicates in all `LastUsedValuesStore`s
We *already* de-duplicate when retrieving values, in all cases where
that's desired, so also de-duplicating when storing values has no impact
on functionality— it's essentially a performance improvement, to reduce
the number of values we store. At 100 values maximum, I don't think this
makes a meaningful difference.

By removing the optimization, we now store history for all values, so if
we choose to enable "smart sort" for other quests in the future, they'll
already have history to pull from.

Also, it simplifies the code, just a little.