Re-type the LastPickedValuesStore
1. With the store being untyped, the compiler does not guarantee that
   the (de)serialization functions match— or that they both use the same
   key to store/retrieve the serialized values.

2. Serialization and deserialization are done relatively far apart.

Together, these make it hard for a human to notice if there's a problem.

Creating a 'Factory' (better name pending) forces the function types to
match, to be declared close together, and to be associated with the same
key. This should make it easy for humans to notice any issues… while
being easier for humans to spot issues *and* more flexible than the
previous typed implementation, where changing the (de)serialization
method required changes to the store.