Fix location pin being based on invisible views
When the current location was underneath a view that was technically
android-visible (View.VISIBLE), it was being shown as if the view was
human-visible. Notably this happened for the undo button and download
indicator, when they were not currently visible.

This is not really a total fix. If a FrameLayout wraps some views of
mixed visibility, the pointer pin will still show at the intersection
with the FrameLayout, rather than just the visible subsections.
Also, it introduces some risk of the pointer pin not being shown when
there is a non-transparent ViewGroup with no children.

Fixes #3460