Fix note form positioning bug
`ViewCompat.setWindowInsetsAnimationCallback` actually hooks in to the
`OnApplyWindowInsetsListener` and runs a little animation (and thus,
calls our callbacks) each time. Among the veritable deluge of animation
callbacks this produces, one (or more) starts before and ends after the
regular `onApplyWindowInsets` callback.

This is not the animation callback we were expecting! Well, maybe we
were expected it, but it is not the keyboard one. So our check to see if
it's an input method animation in `onPrepare` and `onEnd` returns false.

As a result, the animation does not end by applying the latest insets
from `onApplyWindowInsets`; the last effective callback is `onProgress`,
which is not called with the correct insets, for some reason I don't
understand; it might be related to Android internally not marking the
IME as invisible, but I'm not certain.

The previous "manual compat" implementation avoids this by not using the
animation callbacks at all, but this also means the note form doesn't
smoothly move up when the keyboard opens.