Sort building type recents by chronological order
The new "weighted" strategy is good for deciding *which* entries to
show, but it does not provide a stable order; sometimes entries jump
around, which is confusing/disorienting. This returns the sort order to
chronological, while keeping the same strategy for choosing which
entries to show. This means you won't be able to predict *which* entry
will disappear from the list, but you will easily be able to see which
one disappeared after it happens.

More: https://github.com/streetcomplete/StreetComplete/discussions/3493