remove support for "shoulder"
It's usage is negligible (below 1%) and there is a clear overlap with on_street/street_side which will confuse users. A shoulder with parking cars on it will not really be a shoulder anymore in the sense that you can stop there in emergencies etc.

In particular, German translation for "shoulder" is "Seitenstreifen" but according to the wiki, for "Seitenstreifen", i.e. any parking that is adjacent ot the street, street_side should be used.