Tweak amount of answer history considered.
The historyCount value of 30 was picked somewhat arbitrarily, based on a
guess of how much you'd have to consider to have a different result than
just showing the latest 6 values.

For the building levels quest, this was too much, since only 5 values
were shown. When moving between areas with different building heights,
it forced you to re-enter values too many times. Counting only 15 values
should fix this. But we should keep an eye out to make sure we didn't
over-shoot and get rid of any benefit of tracking a larger stretch of
history.

For the building type quest, 30 didn't feel like enough, since the
distribution of answers is different; an item would keep dropping off
the end of the list a few answers before I wanted to use it again.
Bumping it up to 50 fixed that. Testing in a mixed area, it worked
great. But there's a risk that it will cause the same problem as in the
building levels quest; in that case, increasing the `first` value to 2
or 3 is a good next thing to try.