Port manual Canary Island edit again
This ports the manual edit/addition of Canary Island in commit
https://github.com/streetcomplete/StreetComplete/commit/2a6c484ae0e69596a7a2613642ab5c486fc7583c
again.

Unfortunately FAO has no data on that island specificially, so we have no
better data here, so the contribution in that PR should not get lost.

To prevent losing that, I also added an explicit comment, so one does not
have to search the git history again.