doc(contributing): fix 404 link
I did not use a permalink, but kept the link on master branch, because it's likely of interest to see the most up-to-date version of that list.
Occasional breakage is thus okay IMHO. And given this is one of the first lines in that file, even that may not happen that often as before. (i.e. only on file/dir restructure/rename, as the file structure inside the file likely stays as it is)