display street width if set in parking overlay
because sometimes it is quite subjective to decide whether something is still on-street parking or already off-street parking (e.g. parking bays etc.). Since the former counts to the street width and the latter not, this info is somewhat connected with the width - at least the two should not contradict each other