try putting all items, not only in replaced tiles
because the tiles of the "additional" items may already be cached as empty tiles

Example: load 2 empty neighboring tiles into cache, download data in tile 1.
Without this change, tile 2 will still be empty, but we actually have data inside the 20m download-padding in database.
So either this data has to be added to the cache, or the tile should be dropped. Leaving it empty leads to inconsistent behavior, found in some situations.

Also switching back to byTile[item.getTilePos()], as now looping over all tiles in cache would be required, which negates the small performance benefit of the loop variant.