on upload, only broadcast the actually updated element (fixes #4371)
when the OSM API returns the updated element, the uploader fetches other elements from the local db that are necessary to construct the (new) geometry of the updated element (i.e. nodes of a way, member ways+nodes of a relation). Now, these elements have not actually been updated, they are just used to create the geometry. Hence, only the elements returned from the API must be broadcast as having been updated.

The issue identified in #4371 came to pass because that hasn't been the case but at the same time, no geometry for those additional elements was provided and so all the quests referring to elements with apparently missing geometry were removed.