get rid of tracktype overlay
as explained in https://github.com/streetcomplete/StreetComplete/issues/4378#issuecomment-1308396970

StreetComplete now detects string conflicts between surface. And tracktype values are not intuitive, highly abstract and hard to judge in my experience.

I used surface overlay in forest and it was useful (even if not as useful as I expected) but I never used or had need for tracktype overlay.