remove scrollview from time range picker (fixes #4763)
It was added so that the time picker displays... "correctly" (i.e. is not cut off) in landscape mode. However, the time picker and the scrollview do not get along with each other well as #4763 show. So instead, in landscape mode, the time pickers are now shown as spinners. These fit into landscape mode generally.