do not let StyleableOverlayManager add multiple listeners (fixes #4877)
this likely led to memory leaks and thus various crashes because as long as StyleableOverlayManager still listened to the MapDataController, it would not be garbage collected and in turn, any resources it used would also not be garbage collected. This include at least tangram-es MapController which holds at least a reference to Context (the Activity). Hence, the activity would not be garbage collected when it is re-created.