fix bug in filters parser (see #5210#issuecomment-1773483131)
Parser parsed "a or (b or c) and d" into "a or b or c and d". ("and" has higher operator precedence)

It was caused by a premature flattening of the expressions (a or (b or c) => a or b or c) while still parsing the expression. This did not regard that an operator with higher precedence than at the current point in parsing may follow next.