Streamline Flows in LogsViewModelImpl
Previously, the code was not idiomatic to how Kotlin Flows work.
This is corrected with this commit.

To do so, a separate function is used to create a callback flow on the logs controller.
This is then called per new filter, and provided to the UI in that fashion.

In response to work done:
https://github.com/streetcomplete/StreetComplete/pull/5482