make display of quest icons super crisp
quest icons have been optimized for crisp display at 32x32 pixels and multitudes of that. 32x32 seems too small (and 64x64 too big), so let's make it 48x48 which is very close to the previous size of 44x44. The pins are rasterized at double the size so that the selected pins component can zoom in on them.
Also, added a (fake) elevation shadow for the pins and redid the drop shadow.