Clarify wording of the crossing quest (#5936)
* Clarify wording of the crossing quest

I was confused by the wording of the answer "No, but crossing is
possible" to the question "Is there a crossing here?". I needed to check
in the source code of StreetComplete what the answer is saying.

I propose to change the wording. My proposed wording mentions informal
trails and adds a picture of such an informal trail, which would have
helped me.

* scale images correctly

* undo changes in translation strings.xml

* fix images looking weird because of '-colorspace RGB'

* apply suggested wordings

* also bump string name in kotlin code

* revert bumping string