Always move off-rules to the back, regardless where they were specified in the UI (fixes #6191)
Users should ideally write exactly what they see on the sign. Sometimes, opening hours signs start with e.g. Monday off. If then the `off` rule would be written at the start of the generated opening hours string, the other rules would *overwrite* the off-rule according to the OSM opening hours spec. We don't want that, as that's not how the off rules are meant to be interpreted on opening hours signs.

In short, OSM opening hours syntax is not always equal to how shops post their opening hours