avoid integer math in StreetParkingDrawable
With the current images "ic_car*" this is not causing any harm, however if the image
is turned 90° the integer math can lead to 0 as result for height/width. Leading to
Infinity as results.

To avoid issues after exchanging/adding the images in the future, implicit floating
point operations are required.